task 6

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-This module defines a type-annotated function `sum_mixed_list` that calculates 
+This module defines a type-annotated function `sum_mixed_list` that calculates
 the sum of a list containing both integers and floats.
 """
 
@@ -12,7 +12,7 @@ def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     Computes the sum of a list containing both integers and floating-point numbers.
 
     Args:
-        mxd_lst (List[Union[float, int]]): A list of integers and/or floats.
+        mxd_lst (List[Union[int, float]]): A list of integers and/or floats.
 
     Returns:
         float: The sum of the numbers in the input list as a floating-point number.
